// test/src/model/evil/doctoEvil.ts
import { GenerateView, ViewModelType } from "../../../..";
import { BaseEvil } from "./baseEvil";

@GenerateView({
    model: "DoctorEvilViewModel",
    type: 'class',
    filePath: "./test/dist/_autogenerated_viewmodels/evils",
    mapperPath: "./test/dist/_autogenerated_viewmodels/evils/mappers"})
export class DoctorEvil extends BaseEvil {
    hasMiniMe?: boolean;

    firstKilledEvil?: BaseEvil;

    @ViewModelType({type: 'string'})
    notIgnoringFieldInDoctorEvil: boolean;
}